+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+v3.0.316
+----------
+ * Remove stop endpoint for MT
+
+v3.0.315
+----------
+ * Inactive flows should not be listed on the API endpoint
+ * Add Mtarget channel type
+
+v3.0.314
+----------
+ * Add run dict to default webhook payload
+
+v3.0.313
+----------
+ * have URNs resolve to dicts instead of just the display
+ * order transfer credit options by name
+ * show dashboard link even if org is chosen
+
 v3.0.312
 ----------
  * include contact URN in webhook payload